// src/types/generated/Patient.ts
// This file was generated by [ts-rs](https://github.com/Aleph-Alpha/ts-rs). Do not edit this file manually.

export type Patient = { id: number, name: string | null, speciesId: number | null, breedId: number | null, species: string | null, breed: string | null, gender: string | null, dateOfBirth: string | null, color: string | null, weight: number | null, microchipId: string | null, medicalNotes: string | null, isActive: boolean, householdId: number | null, createdAt: string, updatedAt: string, };